feat: AT3-45 Added functional add candidate positions
    Added new column in position table 'sequence' with data type of small integer
    Refactored some codes.
    Server-side is still not the same as client side validation

    Missing deletion of candidate positions
    Missing empty state

had merge conflict on:
    style.css
    core.css
even I had not change those files even the scss partner of those
I have just accepted all incoming changes

style: changing the image and bar thickness based on the size of the screen (#38)

feat: AT1-65 add back-end of feedback and end point page (#39)


chore: added footer, session exchange, sidebar responsiveness

style: just changed the name of the sub-menu

fix: fixed the double quotation marks

bug: fixed the double quotation marks

style: changed the cursor & hover on profile dropdown

style: cursors, active dropdown
