+++ src/includes/model/configuration/endpoint-response.php
@@ -1,0 +1,11 @@
+<?php
+trait EndpointResponse
+{
+    protected static function sendResponse($statusCode, $body, $terminate = false)
+    {
+        http_response_code($statusCode);
+        header('Content-Type: application/json');
+        echo json_encode($body);
+        $terminate && exit;
+    }
+}
