+++ src/includes/organization-list.php
@@ -1,13 +1,50 @@
 <?php
 
 $org_full_names = array(
-    "sco" => "student council organization",
-    "elite" => "eligible league of information technology enthusiasts",
-    "acap" => "association of competent and aspiring psychologists",
-    "aeces" => "association of electronics and communications engineering",
-    "give" => "guild of imporous and valuable educators",
-    "jehra" => "junior executives of human resource association",
-    "jmap" => "junior marketing association of the philippines",
-    "jpia" => "junior philippine institute of accountants",
-    "piie" => "philippine institute of industrial engineers"
+    "sco" => ("student council organization"),
+    "elite" => ("eligible league of information technology enthusiasts"),
+    "acap" => ("association of competent and aspiring psychologists"),
+    "aeces" => ("association of electronics and communications engineering"),
+    "give" => ("guild of imporous and valuable educators"),
+    "jehra" => ("junior executives of human resource association"),
+    "jmap" => ("junior marketing association of the philippines"),
+    "jpia" => ("junior philippine institute of accountants"),
+    "piie" => ("philippine institute of industrial engineers")
+);
+
+$org_acronyms = array(
+    "sco" => ("sco"),
+    "elite" => ("elite"),
+    "acap" => ("acap"),
+    "aeces" => ("aeces"),
+    "give" => ("give"),
+    "jehra" => ("jehra"),
+    "jmap" => ("jmap"),
+    "jpia" => ("jpia"),
+    "piie" => ("piie")
+);
+
+# To be updated
+$org_emails = array(
+    "sco" => ("sco-pup@gmail.com"),
+    "elite" => ("elite-pup@gmail.com"),
+    "acap" => ("acap-pup@gmail.com"),
+    "aeces" => ("aeces-pup@gmail.com"),
+    "give" => ("give-pup@gmail.com"),
+    "jehra" => ("jehra-pup@gmail.com"),
+    "jmap" => ("jmap-pup@gmail.com"),
+    "jpia" => ("jpia-pup@gmail.com"),
+    "piie" => ("piie-pup@gmail.com")
+);
+
+$org_social_media = array(
+    "sco" => array("facebook" => "sco_facebook", "twitter" => "sco_twitter", "instagram" => "sco_instagram"),
+    "elite" => array("facebook" => "elite_facebook", "twitter" => "elite_twitter", "instagram" => "elite_instagram"),
+    "acap" => array("facebook" => "acap_facebook", "twitter" => "acap_twitter", "instagram" => "acap_instagram"),
+    "aeces" => array("facebook" => "aeces_facebook", "twitter" => "aeces_twitter", "instagram" => "aeces_instagram"),
+    "give" => array("facebook" => "give_facebook", "twitter" => "give_twitter", "instagram" => "give_instagram"),
+    "jehra" => array("facebook" => "jehra_facebook", "twitter" => "jehra_twitter", "instagram" => "jehra_instagram"),
+    "jmap" => array("facebook" => "jmap_facebook", "twitter" => "jmap_twitter", "instagram" => "jmap_instagram"),
+    "jpia" => array("facebook" => "jpia_facebook", "twitter" => "jpia_twitter", "instagram" => "jpia_instagram"),
+    "piie" => array("facebook" => "piie_facebook", "twitter" => "piie_twitter", "instagram" => "piie_instagram")
 );
