+++ src/scripts/script.js
@@ -1,28 +1,39 @@
-
 // Sidebar Functionality
 // ----------------------
 
 // Submenu dropdown toggle
-var sidebar = document.querySelector(".sidebar");
-var sidebarClose = document.querySelector("#sidebar-close");
-var menu = document.querySelector(".menu-content");
-var menuItems = document.querySelectorAll(".submenu-item");
-var subMenuTitles = document.querySelectorAll(".submenu .menu-title");
+  var sidebar = document.querySelector(".sidebar");
+  var sidebarClose = document.getElementById("sidebar-close");
 
-sidebarClose.addEventListener("click", () => sidebar.classList.toggle("close"));
+  if (sidebar && sidebarClose) {
+    // Function to toggle sidebar visibility
+    function toggleSidebar() {
+      if (sidebar.classList.contains("open")) {
+        sidebar.classList.remove("open");
+        sidebar.classList.add("close");
+      } else {
+        sidebar.classList.remove("close");
+        sidebar.classList.add("open");
+      }
+    }
 
-menuItems.forEach((item, index) => {
-  item.addEventListener("click", () => {
-    item.classList.toggle("show-submenu");
-    
-    let submenu = item.querySelector('.submenu');
-    $(submenu).collapse('toggle');
-    
-    menuItems.forEach((item2, index2) => {
-      if (index !== index2) {
-        item2.classList.remove("show-submenu");
-        let otherSubmenu = item2.querySelector('.submenu');
-        $(otherSubmenu).collapse('hide'); 
+    sidebarClose.addEventListener("click", toggleSidebar);
+  } else {
+    console.error("Sidebar or sidebar close button not found.");
+  }
+
+
+menuItems.forEach(function (item) {
+  item.addEventListener("click", function () {
+    toggleSidebar();
+
+    var submenu = item.querySelector(".submenu");
+    $(submenu).collapse("toggle");
+
+    menuItems.forEach(function (otherItem) {
+      if (otherItem !== item) {
+        var otherSubmenu = otherItem.querySelector(".submenu");
+        $(otherSubmenu).collapse("hide");
       }
     });
   });
@@ -35,43 +46,44 @@ subMenuTitles.forEach((title) => {
 });
 
 // Changing of toggle icon of submenus
-document.addEventListener('DOMContentLoaded', function () {
-  var submenuToggle = document.getElementById('submenuToggle');
-  var submenuIcon = document.getElementById('submenuIcon');
+document.addEventListener("DOMContentLoaded", function () {
+  var submenuToggle = document.getElementById("submenuToggle");
+  var submenuIcon = document.getElementById("submenuIcon");
 
-  submenuToggle.addEventListener('click', function () {
-      if (submenuIcon.classList.contains('fa-chevron-right')) {
-          submenuIcon.classList.remove('fa-chevron-right');
-          submenuIcon.classList.add('fa-chevron-down');
-      } else {
-          submenuIcon.classList.remove('fa-chevron-down');
-          submenuIcon.classList.add('fa-chevron-right');
-      }
-      submenuIcon.style.transition = 'transform 0.5s ease';
+  submenuToggle.addEventListener("click", function () {
+    if (submenuIcon.classList.contains("fa-chevron-right")) {
+      submenuIcon.classList.remove("fa-chevron-right");
+      submenuIcon.classList.add("fa-chevron-down");
+    } else {
+      submenuIcon.classList.remove("fa-chevron-down");
+      submenuIcon.classList.add("fa-chevron-right");
+    }
+    submenuIcon.style.transition = "transform 0.5s ease";
   });
 });
+
 
 // Checkbox Table Functionality
 // ----------------------
 
-$(document).ready(function(){
-	$('[data-toggle="tooltip"]').tooltip();
-	
-	var checkbox = $('table tbody input[type="checkbox"]');
-	$("#selectAll").click(function(){
-		if(this.checked){
-			checkbox.each(function(){
-				this.checked = true;                        
-			});
-		} else{
-			checkbox.each(function(){
-				this.checked = false;                        
-			});
-		} 
-	});
-	checkbox.click(function(){
-		if(!this.checked){
-			$("#selectAll").prop("checked", false);
-		}
-	});
+$(document).ready(function () {
+  $('[data-toggle="tooltip"]').tooltip();
+
+  var checkbox = $('table tbody input[type="checkbox"]');
+  $("#selectAll").click(function () {
+    if (this.checked) {
+      checkbox.each(function () {
+        this.checked = true;
+      });
+    } else {
+      checkbox.each(function () {
+        this.checked = false;
+      });
+    }
+  });
+  checkbox.click(function () {
+    if (!this.checked) {
+      $("#selectAll").prop("checked", false);
+    }
+  });
 });
