+++ src/styles/candidate-position.css
@@ -21,8 +21,33 @@ button.add-new, main table.table tbody tr {
   height: calc(1rem + 3vw);
 }
 
+body div.modal button.btn {
+  color: var(--white);
+  box-shadow: var(--bs-box-shadow-sm);
+}
+body div.modal button.btn:active {
+  box-shadow: none !important;
+}
+body div.modal button.btn-primary:not(:disabled) {
+  background-color: var(--primary-color);
+  border: calc(0.25rem + 0.2vw) solid var(--primary-color);
+}
+body div.modal button.btn-secondary:not(:disabled) {
+  background-color: var(--secondary-color);
+  border: calc(0.25rem + 0.2vw) solid var(--secondary-color);
+}
 main {
   height: calc(100dvh - 104.95px);
+}
+main button.del {
+  font-weight: bold;
+  position: relative;
+}
+main button.del span.icon.trash {
+  padding: 1rem 0.25rem;
+}
+main button.del span.icon.trash svg {
+  transform: translateY(-8%);
 }
 main table.table {
   margin-bottom: 0px;
