+++ src/styles/core.css
@@ -1,1 +1,83 @@
-:root{--acap: #592293;--aeces: #6610f2;--elite: #EE1C24;--give: #084A6F;--jehra: #A60000;--jmap: #FAD807;--jpia: #F45B9B;--piie: #026644;--sco: #000;--black: #000;--white: #fff;--gray: #6c757d;--gray-dark: #343a40;--primary: #1b00ff;--secondary: #6c757d;--success: #41AE32;--info: #17a2b8;--warning: #ffc107;--danger: #dc3545;--light: #f8f9fa;--dark: #343a40;--breakpoint-xs: 0;--breakpoint-sm: 576px;--breakpoint-md: 768px;--breakpoint-lg: 992px;--breakpoint-xl: 1200px;--font-family-sans-serif: Montserrat, sans-serif;--font-family-monospace: Liberation Mono, 'Courier New', monospace}.mx-w-150{max-width:150px}.mx-h-300{max-height:300px}.mx-h-350{max-height:350px}.mx-h-380{max-height:380px}.mx-h-400{max-height:400px}.mx-h-450{max-height:450px}.pd-0{padding:0px 0px 0px 0px}.padding-top-10{padding:10px 10px 10px 10px}.padding-top-30{padding:30px 30px 30px 30px}.padding-bottom-30{padding:0 0 30px 0}.pd-5{padding:5px 5px 5px 5px}.pd-10{padding:10px 10px 10px 10px}.pd-20{padding:20px 20px 20px 20px}.pd-ltr-20{padding:20px 20px 0 20px 20px 0 20px 20px 0 20px 20px 0}.pd-30{padding:30px 30px 30px 30px}.pb-10{padding:0 0 10px 0}.pb-15{padding:0 0 15px 0}.pb-20{padding:0 0 20px 0}.pb-30{padding:0 0 30px 0}.pt-10{padding:10px 0 10px 0 10px 0 10px 0}.pt-15{padding:15px 0 15px 0 15px 0 15px 0}.pt-20{padding:20px 0 20px 0 20px 0 20px 0}.pt-30{padding:30px 0 30px 0 30px 0 30px 0}.pr-10{padding:0 10px 0 10px}.pr-20{padding:0 20px 0 20px}.pr-30{padding:0 30px 0 30px}.pl-10{padding:0 0 0 10px}.pl-20{padding:0 0 0 20px}.pl-30{padding:0 0 0 30px}.margin-5{margin:5px 5px 5px 5px}.mt-15{margin:15px 15px 15px 15px}.mt-30{margin:30px 30px 30px 30px}.mb-30{margin:0 0 30px 0}.mb-20{margin:0 0 20px 0}.mb-15{margin:0 0 15px 0}.mb-10{margin:0 0 10px 0}.mb-5{margin:0 0 5px 0}.mr-10{margin:0 10px 0 10px}.mr-20{margin:0 20px 0 20px}.font-30{font-size:30px;font-weight:700;line-height:1.3em;color:#000}.font-24{font-size:24px;font-weight:400;line-height:1.42em;color:#000}.font-20{font-size:20px;font-weight:400;line-height:1.5em;color:#000}.font-18{font-size:18px;font-weight:400;line-height:1.65em;color:#000}.font-16{font-size:16px;font-weight:400;line-height:1.62em;color:#000}.font-14{font-size:14px;font-weight:400;line-height:1.71em;color:#000}.font-12{font-size:12px;font-weight:400;line-height:1.83em;color:#000}.h1{font-size:40px;font-weight:700;line-height:1.3;color:#000}.h2{font-size:24px;font-weight:700;line-height:1.35;color:#000}.h3{font-size:20px;font-weight:700;line-height:1.5;color:#000}.h4,.page-header h4{font-size:20px;font-weight:700;line-height:1.35;color:#000}.h5{font-size:18px;font-weight:600;line-height:1.35;color:#000}.line-height-1{line-height:1em}.line-height-1-2{line-height:1.2em}.line-height-1-3{line-height:1.3em}.weight-300{font-weight:300}.weight-400{font-weight:400}.weight-500{font-weight:500}.weight-600{font-weight:600}.weight-700{font-weight:700}.weight-800{font-weight:800}.height-auto{height:auto}.height-100-p{height:100%}.width-50-p{width:50%}.width-box{width:50px;height:50px}.max-width-100{max-width:100px}.max-width-150{max-width:150px}.max-width-200{max-width:200px}.max-width-300{max-width:300px}.max-width-400{max-width:400px}.max-width-500{max-width:500px}.max-width-600{max-width:600px}.max-width-700{max-width:700px}.max-width-800{max-width:800px}.max-width-900{max-width:900px}.max-width-1000{max-width:1000px}.min-height-200px{height:200px}/*# sourceMappingURL=core.css.map */
+:root {
+  --acap: #592293;
+  --aeces: #6610f2;
+  --elite: #ee1c24;
+  --give: #084a6f;
+  --jehra: #a60000;
+  --jmap: #fad807;
+  --jpia: #f45b9b;
+  --piie: #026644;
+  --sco: #000;
+  --main-red: #ea423c;
+  --main-blue: #5d83d1;
+  --black: #000;
+  --white: #fff;
+  --gray: #6c757d;
+  --gray-dark: #4d4d4d;
+  --link-blue: #4169E1;
+  --primary: #1b00ff;
+  --secondary: #6c757d;
+  --success: #41ae32;
+  --info: #17a2b8;
+  --warning: #ffc107;
+  --danger: #dc3545;
+  --light: #f8f9fa;
+  --dark: #343a40;
+  --breakpoint-xs: 0;
+  --breakpoint-sm: 576px;
+  --breakpoint-md: 768px;
+  --breakpoint-lg: 992px;
+  --breakpoint-xl: 1200px;
+  --font-family-sans-serif: Montserrat, sans-serif;
+  --font-family-monospace: Liberation Mono, "Courier New", monospace;
+}
+
+a {
+  text-decoration: none !important;
+}
+
+.main-red {
+  color: #ea423c !important;
+}
+
+.main-blue {
+  color: var(--main-blue) !important;
+}
+
+.dark-gray {
+  color: var(--gray-dark) !important;
+}
+
+.link-blue {
+  color: var(--link-blue);
+}
+
+.spacing-6 {
+  letter-spacing: 0.06rem !important;
+}
+
+.spacing-8 {
+  letter-spacing: 0.08rem !important;
+}
+
+.spacing-10 {
+  letter-spacing: 0.01rem !important;
+}
+
+.spacing-15 {
+  letter-spacing: 0.015rem !important;
+}
+
+fs-12 {
+  font-size: 0.5rem;
+}
+
+.w-35 {
+  width: 35% !important;
+}
+
+.link:hover {
+  color: #4169E1 !important;
+}
+
+/*# sourceMappingURL=core.css.map */
