+++ src/styles/orgs/jehra.css
@@ -2,19 +2,71 @@
 	Color Scheme
 ============================*/
 
-.main-color {
-    color: #A60000;
+:root {
+  --main-color: #a60000;
+  --sub-color-accent: rgb(166, 0, 0, 0.5);
 }
 
-/* Lightest to Darkest */
+.main-color {
+  color: #a60000;
+}
+
+.main-bg-color {
+  background-color: #a60000;
+}
+
+.item a.active,
+.submenu-item.active {
+  background-color: #a60000;
+  color: white;
+}
+
+.item a.active:hover,
+.submenu-item.active:hover {
+  background-color: #a60000;
+  color: white;
+}
+
+.org-info {
+  background-color: #a60000;
+}
+
+.org-logo {
+  height: 100px;
+  border-radius: 50%;
+  margin-right: 10px;
+  border: 2px solid #a60000;
+}
+
+.item a:hover {
+  background: rgb(166, 0, 0, 0.1);
+}
+
+.submenu-item:hover {
+  background: rgb(166, 0, 0, 0.1);
+}
+
+.submenu .menu-title:hover {
+  background: rgb(166, 0, 0, 0.1);
+}
+
+.menu-items .item a.active-sub {
+  color: #a60000;
+}
+
+
+/*==========================
+	ACCENT COLORS: L TO D
+============================*/
+
 .accent-1 {
-    color: #C20E31;
+  color: #c20e31;
 }
 
 .accent-2 {
-    color: #A60000;
+  color: #a60000;
 }
 
 .accent-3 {
-    color: #81171B;
+  color: #81171b;
 }
