--- src/styles/scss/_button-del.scss
@@ -1,24 +1,0 @@
-main.main button {
-
-    &.btn {
-        color: var(--white);
-        box-shadow: var(--bs-box-shadow-sm);
-
-        &:active {
-            box-shadow: none !important;
-        }
-    }
-
-    &.del {
-        font-weight: bold;
-        position: relative;
-
-        span.icon.trash {
-            padding: 1rem 0.25rem;
-
-            svg {
-                transform: translateY(-8%);
-            }
-        }
-    }
-}
