+++ src/styles/scss/candidate-position.scss
@@ -36,6 +36,44 @@
     height: calc(1rem + 3vw);
 }
 
+body div.modal {
+    button {
+
+        &.btn {
+            color: var(--white);
+            box-shadow: var(--bs-box-shadow-sm);
+
+            &:active {
+                box-shadow: none !important;
+            }
+        }
+
+        &.btn-primary:not(:disabled) {
+            background-color: var(--primary-color);
+            border: calc(0.25rem + 0.2vw) solid var(--primary-color);
+
+            &:hover {}
+
+            &:active {}
+
+            &:focus {}
+        }
+
+        &.btn-secondary:not(:disabled) {
+            background-color: var(--secondary-color);
+            border: calc(0.25rem + 0.2vw) solid var(--secondary-color);
+
+            &:hover {}
+
+            &:active {}
+
+            &:focus {}
+        }
+
+
+    }
+}
+
 main {
 
     height: calc(100dvh - (58.96px + 45.99px));
@@ -49,6 +87,19 @@ main {
     }
 
     @extend %input-text-editable;
+
+    & button.del {
+        font-weight: bold;
+        position: relative;
+
+        span.icon.trash {
+            padding: 1rem 0.25rem;
+
+            svg {
+                transform: translateY(-8%);
+            }
+        }
+    }
 
     table.table {
         margin-bottom: 0px;
